fix a memory accounting overflow when using memacctd for reporting on memory usage
git-svn-id: svn://clusterresources.com/torque/branches/3.0-fixes@4614 8f940c70-5916-0410-ac46-d1fa2fa6ea29

+++ src/resmom/linux/mom_mach.c
@@ -216,7 +216,7 @@ static char *walltime (struct rm_attribute *);
 static char *quota    (struct rm_attribute *);
 static char *netload  (struct rm_attribute *);
 #ifdef USELIBMEMACCT
-static long get_memacct_resi(pid_t pid);
+static long long get_memacct_resi(pid_t pid);
 extern long get_weighted_memory_size(pid_t);
 #endif
 
@@ -471,10 +471,10 @@ proc_stat_t *get_proc_stat(
  * Returns the value in bytes on success, returns -1 on failure.
  */
 
-long get_memacct_resi(pid_t pid)
+long long get_memacct_resi(pid_t pid)
   {
   char *id = "get_memacct_resi";
-  long  w_rss;
+  long long  w_rss;
 
   if ((w_rss = get_weighted_memory_size(pid)) == -1)
     {
@@ -1135,7 +1135,7 @@ static unsigned long long resi_sum(
   unsigned long long  resisize;
   proc_stat_t        *ps;
 #ifdef USELIBMEMACCT
-  long                w_rss;
+  long long                w_rss;
 #endif
 
   resisize = 0;
@@ -3122,7 +3122,7 @@ static char *resi_job(
   unsigned long long  resisize;
   proc_stat_t        *ps;
 #ifdef USELIBMEMACCT
-  long                w_rss;
+  long long           w_rss;
 #endif
 
   resisize = 0;
@@ -3192,7 +3192,7 @@ static char *resi_proc(
   proc_stat_t *ps;
 
 #ifdef USELIBMEMACCT
-  long          w_rss;
+  long long     w_rss;
 #endif
 
   if ((ps = get_proc_stat(pid)) == NULL)
